Fix os depended definition
